Fix for infinite looping in poll mode
The _poll() method only gets succesfully called only if an error did
not happened during polling.
When Telegram throws a 409 error the bot start increasingly making
setTimeout calls that enqueue .poll() methods that span the CPU up to
100% rapidly.
This way we leave the developer to handle the error code and then poll
again if they need to.